link logic
