--- apps/expo/src/components/SignInScreen.tsx
@@ -1,59 +1,0 @@
-import { useState } from "react";
-import { Text, TextInput, TouchableOpacity, View } from "react-native";
-import Toast from "react-native-root-toast";
-import { useSignIn } from "@clerk/clerk-expo";
-
-export default function SignInScreen() {
-  const { signIn, setActive, isLoaded } = useSignIn();
-
-  const [emailAddress, setEmailAddress] = useState("");
-  const [password, setPassword] = useState("");
-
-  const onSignInPress = async () => {
-    if (!isLoaded) {
-      return;
-    }
-
-    try {
-      const completeSignIn = await signIn.create({
-        identifier: emailAddress,
-        password,
-      });
-      // This is an important step,
-      // This indicates the user is signed in
-      await setActive({ session: completeSignIn.createdSessionId });
-    } catch (err) {
-      const errorMessage = `Error signing in: ${
-        err instanceof Error ? err.message : "unknown error"
-      }`;
-      Toast.show(errorMessage, {
-        duration: Toast.durations.LONG,
-      });
-    }
-  };
-  return (
-    <View>
-      <View>
-        <TextInput
-          autoCapitalize="none"
-          value={emailAddress}
-          placeholder="Email..."
-          onChangeText={(emailAddress) => setEmailAddress(emailAddress)}
-        />
-      </View>
-
-      <View>
-        <TextInput
-          value={password}
-          placeholder="Password..."
-          secureTextEntry={true}
-          onChangeText={(password) => setPassword(password)}
-        />
-      </View>
-
-      <TouchableOpacity onPress={onSignInPress}>
-        <Text>Sign in</Text>
-      </TouchableOpacity>
-    </View>
-  );
-}
