+++ apps/expo/src/interface/IClerkError.ts
@@ -1,0 +1,12 @@
+export default interface IClerkError {
+  status: number;
+  clerkError: boolean;
+  errors: {
+    code: string;
+    message: string;
+    longMessage: string;
+    meta: {
+      paramName: string;
+    };
+  }[];
+}
